+++ packages/storybook/stories/CustomIcons/CustomIcons.stories.mdx
@@ -1,6 +1,6 @@
 import { DEFAULT_COMPONENT_ICONS, DEFAULT_COMPOSITE_ICONS } from '@azure/communication-react';
 import { Meta, Canvas, Source } from '@storybook/addon-docs';
-import { IconGridWithSearch } from './utils';
+import { IconGridWithSearch } from './IconGrid';
 
 <Meta id="custom-icons" title="Concepts/Custom Icons" />
